Uncouple Xtext from SADL
Remove Xtext from SADL's feature.xml since we do not want SADL to
require a specific version of Xtext.  SADL should be able to work with
any recent or future version of Xtext as long as no one makes a
breaking change to the Xtext API used by SADL.  Now you can upgrade
your Eclipse DSL/SADL installation to 2021-06 without getting a
conflict saying that SADL requires Xtext 2.24.0 to be installed which
is incompatible with Xtext 2.25.0.