Let SADL run on Java 11
Keep building SADL with Java 17, but allow SADL to run on Java 11.  We
have to use Java 17 to build SADL with Tycho, but we want SADL to work
on both Java 11 and Java 17 at runtime.

**/META-INF/MANIFEST.MF: Roll back Bundle-RequiredExecutionEnvironment
from JavaSE-17 to JavaSE-11.

pom.xml: Allow SADL classes to run on Java 11 instead of requiring Java 17.